+++ src/components/pages/profile/[name]/tabs/AdvancedTab/AdvancedTab.tsx
