+++ src/components/pages/profile/[name]/tabs/AdvancedTab/Fuses.test.tsx
