+++ src/components/pages/profile/[name]/tabs/AdvancedTab/ResolverDetails/ResolverDetails.tsx
