+++ src/components/pages/profile/[name]/tabs/DetailSnippet.test.tsx
