+++ src/components/pages/profile/[name]/tabs/DetailSnippet.tsx
