+++ src/components/pages/profile/[name]/tabs/details.test.tsx
@@ -7,9 +7,9 @@ import { useAccount } from 'wagmi'
 import { NFTWithPlaceholder } from '@app/components/NFTWithPlaceholder'
 import { NameSnippetMobile } from '@app/components/pages/profile/NameSnippetMobile'
 import { OwnerButton } from '@app/components/pages/profile/OwnerButton'
-import More from '@app/components/pages/profile/[name]/details/AdvancedTab/AdvancedTab'
-import { RecordsTab } from '@app/components/pages/profile/[name]/details/RecordsTab'
-import { SubnamesTab } from '@app/components/pages/profile/[name]/details/SubnamesTab'
+import More from '@app/components/pages/profile/[name]/tabs/AdvancedTab/AdvancedTab'
+import { RecordsTab } from '@app/components/pages/profile/[name]/tabs/RecordsTab'
+import { SubnamesTab } from '@app/components/pages/profile/[name]/tabs/SubnamesTab'
 import { useChainId } from '@app/hooks/useChainId'
 import { useNameDetails } from '@app/hooks/useNameDetails'
 import Page, { Details } from '@app/pages/profile/details'
